backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
psycopg2-binary
python-dotenv
pillow